:dizzy:[update]friendList取得外部化
